rename: package path modification
