+++ pennyway-socket/src/main/java/kr/co/pennyway/socket/relay/ChatJoinEventListener.java
@@ -2,8 +2,8 @@ package kr.co.pennyway.socket.relay;
 
 import kr.co.pennyway.infra.common.event.ChatRoomJoinEvent;
 import kr.co.pennyway.infra.common.properties.ChatExchangeProperties;
+import kr.co.pennyway.socket.command.SendMessageCommand;
 import kr.co.pennyway.socket.common.constants.SystemMessageTemplate;
-import kr.co.pennyway.socket.dto.SendMessageCommand;
 import kr.co.pennyway.socket.service.ChatMessageSendService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
